Code is not yet functional, Config file parse is almost complete, Still have to handle quoted strings + optimize and remove repetetive blocks of code + test

+++ helper_functions.cpp
@@ -6,7 +6,7 @@
 /*   By: ymafaman <ymafaman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/11 21:38:05 by ymafaman          #+#    #+#             */
-/*   Updated: 2024/10/11 21:41:10 by ymafaman         ###   ########.fr       */
+/*   Updated: 2024/10/12 15:19:38 by ymafaman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 bool is_all_digits(const std::string& str)
 {
-    for (int i = 0; i < str.length(); i++)
+    for (size_t i = 0; i < str.length(); i++)
     {
         if (!std::isdigit(str[i]))
         {
